增加账户等菜单

+++ publiccms-parent/publiccms-trade/src/main/java/com/publiccms/views/directive/trade/TradeOrderDirective.java
@@ -26,11 +26,11 @@ public class TradeOrderDirective extends AbstractTemplateDirective {
 
     @Override
     public void execute(RenderHandler handler) throws IOException, Exception {
-        Integer id = handler.getInteger("id");
+        Long id = handler.getLong("id");
         if (CommonUtils.notEmpty(id)) {
             handler.put("object", service.getEntity(id)).render();
         } else {
-            Integer[] ids = handler.getIntegerArray("ids");
+            Long[] ids = handler.getLongArray("ids");
             if (CommonUtils.notEmpty(ids)) {
                 List<TradeOrder> entityList = service.getEntitys(ids);
                 Map<String, TradeOrder> map = new LinkedHashMap<>();
